task: Add recommended fields to the Educational converter
